termine su implementacion de los constructores

+++ src/Entidades/PowerUp/Fruta.java
@@ -11,8 +11,8 @@ public class Fruta extends PowerUp{
     protected int puntaje;
     protected float tiempoDeVida;
 
-    public Fruta(Skin s, ModoDeJuego juego, float x, float y, Hitbox h) {
-        super(s, juego, (int) x,(int)y, 300, 10);
+    public Fruta(Skin s, ModoDeJuego juego, int x, int y, Hitbox h) {
+        super(s, juego, x, y, 300, 10);
         puntaje = 500;
         tiempoDeVida = 10;
     }
